refactor(shared): split dynamic truncator facade
